pushing metric_list changes
