+++ great_expectations/experimental/metric_repository/metric_list_metric_retriever.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from itertools import chain
-from typing import TYPE_CHECKING, Any, List, Optional, Sequence
+from typing import TYPE_CHECKING, List, Optional, Sequence
 
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.experimental.metric_repository.metric_retriever import (
@@ -17,9 +17,6 @@ from great_expectations.experimental.metric_repository.metrics import (
 if TYPE_CHECKING:
     from great_expectations.data_context import AbstractDataContext
     from great_expectations.datasource.fluent.batch_request import BatchRequest
-    from great_expectations.validator.metrics_calculator import (
-        _MetricKey,
-    )
     from great_expectations.validator.validator import (
         Validator,
     )
@@ -43,7 +40,7 @@ class MetricListMetricRetriever(MetricRetriever):
 
         self._check_valid_metric_types(metric_list)
 
-        table_metrics = self._get_table_metrics(
+        table_metrics = self._calculate_table_metrics(
             batch_request=batch_request, metric_list=metric_list
         )
         metrics_result.extend(table_metrics)
@@ -116,38 +113,13 @@ class MetricListMetricRetriever(MetricRetriever):
 
         if not metrics_to_calculate:
             return metrics
-
-        column_metric_configs = self._generate_column_metric_configurations(
-            column_list, list(metrics_to_calculate)
-        )
-        batch_id, computed_metrics, aborted_metrics = self._compute_metrics(
-            batch_request, column_metric_configs
-        )
-
-        # Convert computed_metrics
-        ColumnMetric.update_forward_refs()
-        metric_lookup_key: _MetricKey
-
-        for metric_name in metrics_to_calculate:
-            for column in column_list:
-                metric_lookup_key = (metric_name, f"column={column}", tuple())
-                value, exception = self._get_metric_from_computed_metrics(
-                    metric_name=metric_name,
-                    metric_lookup_key=metric_lookup_key,
-                    computed_metrics=computed_metrics,
-                    aborted_metrics=aborted_metrics,
-                )
-                metrics.append(
-                    ColumnMetric[int](
-                        batch_id=batch_id,
-                        metric_name=metric_name,
-                        column=column,
-                        value=value,
-                        exception=exception,
-                    )
-                )
-
-        return metrics
+        else:
+            return self._get_column_metrics(
+                batch_request=batch_request,
+                column_list=column_list,
+                column_metric_names=list(metrics_to_calculate),
+                column_metric_type=ColumnMetric[int],
+            )
 
     def _get_numeric_column_metrics(
         self,
@@ -225,10 +197,10 @@ class MetricListMetricRetriever(MetricRetriever):
             column_metric_type=ColumnMetric[str],
         )
 
-    def _get_table_metrics(
+    def _calculate_table_metrics(
         self, batch_request: BatchRequest, metric_list: List[MetricTypes]
     ) -> List[Metric]:
-        """Calculate column metrics for table metrics, which include row_count, column names and types.
+        """Calculate table metrics, which include row_count, column names and types.
 
         Args:
             metrics_list (List[MetricTypes]): list of metrics sent from Agent.
@@ -255,23 +227,10 @@ class MetricListMetricRetriever(MetricRetriever):
         Returns:
             Metric: Row count for the table.
         """
-        table_metric_configs = self._generate_table_metric_configurations(
-            table_metric_names=[MetricTypes.TABLE_ROW_COUNT]
-        )
-        batch_id, computed_metrics, aborted_metrics = self._compute_metrics(
-            batch_request, table_metric_configs
-        )
-        metric_name = MetricTypes.TABLE_ROW_COUNT
-        value, exception = self._get_metric_from_computed_metrics(
-            metric_name=metric_name,
-            computed_metrics=computed_metrics,
-            aborted_metrics=aborted_metrics,
-        )
-        return TableMetric[int](
-            batch_id=batch_id,
-            metric_name=metric_name,
-            value=value,
-            exception=exception,
+        return self._get_table_metrics(
+            batch_request=batch_request,
+            metric_name=MetricTypes.TABLE_ROW_COUNT,
+            metric_type=TableMetric[int],
         )
 
     def _get_table_columns(self, batch_request: BatchRequest) -> Metric:
@@ -283,23 +242,10 @@ class MetricListMetricRetriever(MetricRetriever):
         Returns:
             Metric: Column names for the table.
         """
-        table_metric_configs = self._generate_table_metric_configurations(
-            table_metric_names=[MetricTypes.TABLE_COLUMNS]
-        )
-        batch_id, computed_metrics, aborted_metrics = self._compute_metrics(
-            batch_request, table_metric_configs
-        )
-        metric_name = MetricTypes.TABLE_COLUMNS
-        value, exception = self._get_metric_from_computed_metrics(
-            metric_name=metric_name,
-            computed_metrics=computed_metrics,
-            aborted_metrics=aborted_metrics,
-        )
-        return TableMetric[List[str]](
-            batch_id=batch_id,
-            metric_name=metric_name,
-            value=value,
-            exception=exception,
+        return self._get_table_metrics(
+            batch_request=batch_request,
+            metric_name=MetricTypes.TABLE_COLUMNS,
+            metric_type=TableMetric[List[str]],
         )
 
     def _get_table_column_types(self, batch_request: BatchRequest) -> Metric:
@@ -311,39 +257,10 @@ class MetricListMetricRetriever(MetricRetriever):
         Returns:
             Metric: Column types for the table.
         """
-        table_metric_configs = self._generate_table_metric_configurations(
-            table_metric_names=[MetricTypes.TABLE_COLUMN_TYPES]
-        )
-        batch_id, computed_metrics, aborted_metrics = self._compute_metrics(
-            batch_request, table_metric_configs
-        )
-        metric_name = MetricTypes.TABLE_COLUMN_TYPES
-        metric_lookup_key: _MetricKey = (metric_name, tuple(), "include_nested=True")
-        value, exception = self._get_metric_from_computed_metrics(
-            metric_name=metric_name,
-            metric_lookup_key=metric_lookup_key,
-            computed_metrics=computed_metrics,
-            aborted_metrics=aborted_metrics,
-        )
-        raw_column_types: list[dict[str, Any]] = value
-        # If type is not found, don't add empty type field. This can happen if our db introspection fails.
-        column_types_converted_to_str: list[dict[str, str]] = []
-        for raw_column_type in raw_column_types:
-            if raw_column_type.get("type"):
-                column_types_converted_to_str.append(
-                    {
-                        "name": raw_column_type["name"],
-                        "type": str(raw_column_type["type"]),
-                    }
-                )
-            else:
-                column_types_converted_to_str.append({"name": raw_column_type["name"]})
-
-        return TableMetric[List[str]](
-            batch_id=batch_id,
-            metric_name=metric_name,
-            value=column_types_converted_to_str,
-            exception=exception,
+        return self._get_table_metrics_column_types(
+            batch_request=batch_request,
+            metric_name=MetricTypes.TABLE_COLUMN_TYPES,
+            metric_type=TableMetric[List[str]],
         )
 
     def _check_valid_metric_types(self, metric_list: List[MetricTypes]) -> bool:
