+++ great_expectations/experimental/metric_repository/metric_retriever.py
@@ -174,6 +174,65 @@ class MetricRetriever(abc.ABC):
         )
         return column_names
 
+    def _get_table_metrics(
+        self,
+        batch_request: BatchRequest,
+        metric_name: MetricTypes | str,
+        metric_type: type[Metric],
+    ) -> Metric:
+        metric_configs = self._generate_table_metric_configurations([metric_name])
+        batch_id, computed_metrics, aborted_metrics = self._compute_metrics(
+            batch_request, metric_configs
+        )
+        value, exception = self._get_metric_from_computed_metrics(
+            metric_name=metric_name,
+            computed_metrics=computed_metrics,
+            aborted_metrics=aborted_metrics,
+        )
+        return metric_type(
+            batch_id=batch_id, metric_name=metric_name, value=value, exception=exception
+        )
+
+    def _get_table_metrics_column_types(
+        self,
+        batch_request: BatchRequest,
+        metric_name: MetricTypes | str,
+        metric_type: type[Metric],
+    ) -> Metric:
+        metric_lookup_key: _MetricKey = (metric_name, tuple(), "include_nested=True")
+        table_metric_configs = self._generate_table_metric_configurations(
+            table_metric_names=[MetricTypes.TABLE_COLUMN_TYPES]
+        )
+        batch_id, computed_metrics, aborted_metrics = self._compute_metrics(
+            batch_request, table_metric_configs
+        )
+        value, exception = self._get_metric_from_computed_metrics(
+            metric_name=metric_name,
+            metric_lookup_key=metric_lookup_key,
+            computed_metrics=computed_metrics,
+            aborted_metrics=aborted_metrics,
+        )
+        raw_column_types: list[dict[str, Any]] = value
+        # If type is not found, don't add empty type field. This can happen if our db introspection fails.
+        column_types_converted_to_str: list[dict[str, str]] = []
+        for raw_column_type in raw_column_types:
+            if raw_column_type.get("type"):
+                column_types_converted_to_str.append(
+                    {
+                        "name": raw_column_type["name"],
+                        "type": str(raw_column_type["type"]),
+                    }
+                )
+            else:
+                column_types_converted_to_str.append({"name": raw_column_type["name"]})
+
+        return metric_type(
+            batch_id=batch_id,
+            metric_name=metric_name,
+            value=column_types_converted_to_str,
+            exception=exception,
+        )
+
     def _get_column_metrics(
         self,
         batch_request: BatchRequest,
@@ -189,6 +248,7 @@ class MetricRetriever(abc.ABC):
         )
 
         # Convert computed_metrics
+        ColumnMetric.update_forward_refs()
         metrics: list[Metric] = []
         metric_lookup_key: _MetricKey
